// apps/k-pay-app/src/types/graphql.ts
export interface PersonalInfoInput {
  firstName: string;
  lastName: string;
  maidenName?: string;
  dateOfBirth: string;
  placeOfBirth: string;
  nationality: string;
  countryOrTaxResidence: string;
  taxIdentificationNumber: string;
  occupation: string;
  currentEmployer: string;
  employmentStatus: string;
}